Driver: I2C: STM32F2/STM32F4/STM32L1: Fix alternate i2c read.

+++ targets/TARGET_STM/TARGET_STM32F4/objects.h
@@ -106,7 +106,7 @@ struct i2c_s {
     int scl_func;
     IRQn_Type event_i2cIRQ;
     IRQn_Type error_i2cIRQ;
-    uint8_t XferOperation;
+    uint32_t XferOperation;
     volatile uint8_t event;
 #if DEVICE_I2CSLAVE
     uint8_t slave;
